+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/addons/InstalledAddonDetailsFragment.kt
@@ -106,26 +106,22 @@ class InstalledAddonDetailsFragment : Fragment() {
             // Only needed in case we are not able to find the add-on.
             var breadcrumb: Breadcrumb? = null
             try {
-                val addons = provideAddonManager().getAddons()
+                val latestAddon = provideAddonManager().getAddonByID(addon.id)
                 runIfFragmentIsAttached {
-                    addons.find { addon.id == it.id }.let {
-                        if (it == null) {
-                            val addonsStringList = addons.joinToString { item -> item.id }
-                            breadcrumb = Breadcrumb(
-                                "Addon ${addon.id} not found, isInstalled: ${addon.isInstalled()}," +
-                                    " add-ons: $addonsStringList",
-                            )
-                            throw AddonManagerException(Exception("Addon ${addon.id} not found"))
-                        } else {
-                            withContext(Dispatchers.Main) {
-                                addon = it
-                                bindUI()
-                            }
-                        }
+                    if (latestAddon == null) {
+                        breadcrumb = Breadcrumb(
+                            "Addon ${addon.id} not found, isInstalled: ${addon.isInstalled()}",
+                        )
+                        throw AddonManagerException(Exception("Addon ${addon.id} not found"))
+                    } else {
                         withContext(Dispatchers.Main) {
-                            binding.addOnProgressBar.isVisible = false
-                            binding.addonContainer.isVisible = true
+                            addon = latestAddon
+                            bindUI()
                         }
+                    }
+                    withContext(Dispatchers.Main) {
+                        binding.addOnProgressBar.isVisible = false
+                        binding.addonContainer.isVisible = true
                     }
                 }
             } catch (e: AddonManagerException) {